refactor new virtualbox service
