+++ core/core.go
@@ -55,8 +55,7 @@ type SubnetNode struct {
 	Firewall        firewall.FirewallInterface    `optional:"true"`
 	Deployer        *deployer.Service             `optional:"true"`
 	K8sDeployer     k8s.Service                   `optional:"true"`
-	VirtualBox      *virtualbox.VirtualboxService `optional:"true"`
-	VirtualBoxV2    virtualbox.IVirtualboxService `optional:"true"`
+	VirtualBox      virtualbox.IVirtualboxService `optional:"true"`
 
 	Account *account.AccountService `optional:"true"`
 	Peers   *peers.Service          `optional:"true"`
