+++ core/virtualbox/storage/image-storage.go
