+++ core/virtualbox/storage/image-storage_test.go
