+++ core/virtualbox/storage/ubuntu-img-url-generator.go
@@ -5,79 +5,65 @@ import (
 	"io"
 	"net/http"
 	"regexp"
+	"runtime"
 	"sort"
 	"strings"
 	"time"
 )
 
 type ImageRequest struct {
-	OS      string `json:"Os"`
-	Version string `json:"Version"`
+	OS      string `json:"os"`
+	Version string `json:"version"`
 }
 
-type UbuntuImageGenerator struct {
-	versionToCodename map[string]string
-	client            *http.Client
+var ubuntuVersionToCodename = map[string]string{
+	"24.04": "noble",
+	"22.04": "jammy",
+	"20.04": "focal",
+	"18.04": "bionic",
+	"16.04": "xenial",
+	"14.04": "trusty",
+	"23.10": "mantic",
+	"23.04": "lunar",
+	"22.10": "kinetic",
+	"21.10": "impish",
+	"21.04": "hirsute",
+	"20.10": "groovy",
+	"19.10": "eoan",
+	"19.04": "disco",
+	"18.10": "cosmic",
 }
 
-func NewUbuntuImageGenerator() *UbuntuImageGenerator {
-	return &UbuntuImageGenerator{
-		versionToCodename: map[string]string{
-			"24.04": "noble",
-			"22.04": "jammy",
-			"20.04": "focal",
-			"18.04": "bionic",
-			"16.04": "xenial",
-			"14.04": "trusty",
-			"23.10": "mantic",
-			"23.04": "lunar",
-			"22.10": "kinetic",
-			"21.10": "impish",
-			"21.04": "hirsute",
-			"20.10": "groovy",
-			"19.10": "eoan",
-			"19.04": "disco",
-			"18.10": "cosmic",
-		},
-		client: &http.Client{
-			Timeout: 10 * time.Second,
-		},
-	}
-}
+func (is *ImageStorage) GenerateImageURL(req ImageRequest) (string, error) {
 
-// GenerateImageURL generates the appropriate Ubuntu cloud image URL
-func (u *UbuntuImageGenerator) GenerateImageURL(req ImageRequest, arch string) (string, error) {
 	if strings.ToLower(req.OS) != "ubuntu" {
 		return "", fmt.Errorf("unsupported OS: %s", req.OS)
 	}
 
-	codename, exists := u.versionToCodename[req.Version]
+	codename, exists := ubuntuVersionToCodename[req.Version]
 	if !exists {
 		return "", fmt.Errorf("unsupported Ubuntu version: %s", req.Version)
 	}
 
+	arch := runtime.GOARCH
 	if arch == "" {
-		arch = "amd64" // default architecture
+		arch = "amd64"
 	}
 
-	// Get the latest available date for this codename
-	latestDate, err := u.getLatestDate(codename)
+	latestDate, err := is.getLatestDate(codename)
 	if err != nil {
-		return "", fmt.Errorf("failed to get latest date for %s: %v", codename, err)
+		return "", err
 	}
 
-	// Construct the URL
-	url := fmt.Sprintf("https://cloud-images.ubuntu.com/%s/%s/%s-server-cloudimg-%s.img",
-		codename, latestDate, codename, arch)
-
-	return url, nil
+	return fmt.Sprintf("https://cloud-images.ubuntu.com/%s/%s/%s-server-cloudimg-%s.img",
+		codename, latestDate, codename, arch), nil
 }
 
 // getLatestDate fetches the directory listing and finds the latest date
-func (u *UbuntuImageGenerator) getLatestDate(codename string) (string, error) {
+func (is *ImageStorage) getLatestDate(codename string) (string, error) {
 	url := fmt.Sprintf("https://cloud-images.ubuntu.com/%s/", codename)
 
-	resp, err := u.client.Get(url)
+	resp, err := is.httpClient.Get(url)
 	if err != nil {
 		return "", err
 	}
@@ -93,7 +79,7 @@ func (u *UbuntuImageGenerator) getLatestDate(codename string) (string, error) {
 	}
 
 	// Parse HTML to find date directories
-	dates := u.parseDateDirectories(string(body))
+	dates := is.parseDateDirectories(string(body))
 	if len(dates) == 0 {
 		return "", fmt.Errorf("no date directories found")
 	}
@@ -104,7 +90,7 @@ func (u *UbuntuImageGenerator) getLatestDate(codename string) (string, error) {
 }
 
 // parseDateDirectories extracts date directories from HTML listing
-func (u *UbuntuImageGenerator) parseDateDirectories(html string) []string {
+func (is *ImageStorage) parseDateDirectories(html string) []string {
 	// Match date patterns like "20250626/" in href attributes
 	datePattern := regexp.MustCompile(`href="(\d{8})/"`)
 	matches := datePattern.FindAllStringSubmatch(html, -1)
@@ -126,23 +112,18 @@ func (u *UbuntuImageGenerator) parseDateDirectories(html string) []string {
 }
 
 // GenerateImageURLWithFallback generates URL with fallback to a known good date
-func (u *UbuntuImageGenerator) GenerateImageURLWithFallback(req ImageRequest, arch string) string {
-	url, err := u.GenerateImageURL(req, arch)
+func (is *ImageStorage) GenerateImageURLWithFallback(req ImageRequest) string {
+	url, err := is.GenerateImageURL(req)
 	if err != nil {
 		// Fallback to current date format if API fails
-		codename, exists := u.versionToCodename[req.Version]
+		codename, exists := ubuntuVersionToCodename[req.Version]
 		if !exists {
 			return ""
 		}
 
-		if arch == "" {
-			arch = "amd64"
-		}
-
-		// Use current date as fallback
-		currentDate := time.Now().Format("20060102")
+		arch := runtime.GOARCH
 		url = fmt.Sprintf("https://cloud-images.ubuntu.com/%s/%s/%s-server-cloudimg-%s.img",
-			codename, currentDate, codename, arch)
+			codename, time.Now().Format("20060102"), codename, arch)
 	}
 	return url
 }
